+++ lib/command_palette.dart
@@ -8,6 +8,7 @@ import 'src/models/command_palette_style.dart';
 export 'src/command_palette.dart';
 export 'src/models/command_palette_action.dart';
 export 'src/models/command_palette_style.dart';
+export 'src/models/command_palette_config.dart';
 
 /// Defines the type of function to be used for filtering command palette
 /// actions.